+++ src/biz/casbin/access_control.rs
@@ -80,7 +80,7 @@ impl CasbinAccessControl {
     }
   }
 
-  async fn update(
+  pub async fn update(
     &self,
     uid: &i64,
     obj: &ObjectType<'_>,
@@ -89,7 +89,7 @@ impl CasbinAccessControl {
     enforcer_update(&self.enforcer, uid, obj, act).await
   }
 
-  async fn remove(&self, uid: &i64, obj: &ObjectType<'_>) -> Result<bool, AppError> {
+  pub async fn remove(&self, uid: &i64, obj: &ObjectType<'_>) -> Result<bool, AppError> {
     enforcer_remove(&self.enforcer, uid, obj).await
   }
 
@@ -424,594 +424,6 @@ impl WorkspaceAccessControl for CasbinWorkspaceAccessControl {
       .casbin_access_control
       .remove(uid, &ObjectType::Workspace(&workspace_id.to_string()))
       .await?;
-    Ok(())
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use anyhow::Context;
-  use casbin::{DefaultModel, Enforcer};
-  use shared_entity::dto::workspace_dto::{CreateWorkspaceMember, WorkspaceMemberChangeset};
-
-  use crate::biz::{
-    self,
-    casbin::{adapter::PgAdapter, tests},
-    pg_listener::PgListeners,
-  };
-
-  use super::*;
-
-  #[sqlx::test(migrations = false)]
-  async fn test_workspace_access_control_get_role(pool: PgPool) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-    let access_control = access_control.new_workspace_access_control();
-
-    let user = tests::create_user(&pool).await?;
-
-    // Get workspace details
-    let workspace = database::workspace::select_user_workspace(&pool, &user.uuid)
-      .await?
-      .into_iter()
-      .next()
-      .ok_or(anyhow!("workspace should be created"))?;
-
-    assert_eq!(
-      AFRole::Owner,
-      access_control
-        .get_role_from_uuid(&user.uuid, &workspace.workspace_id)
-        .await?
-    );
-
-    assert_eq!(
-      AFRole::Owner,
-      access_control
-        .get_role_from_uid(&user.uid, &workspace.workspace_id)
-        .await?
-    );
-
-    let member = tests::create_user(&pool).await?;
-    let _ = biz::workspace::ops::add_workspace_members(
-      &pool,
-      &member.uuid,
-      &workspace.workspace_id,
-      vec![CreateWorkspaceMember {
-        email: member.email.clone(),
-        role: AFRole::Member,
-      }],
-    )
-    .await
-    .context("adding users to workspace")?;
-
-    assert_eq!(
-      AFRole::Member,
-      access_control
-        .get_role_from_uuid(&member.uuid, &workspace.workspace_id)
-        .await?
-    );
-
-    assert_eq!(
-      AFRole::Member,
-      access_control
-        .get_role_from_uid(&member.uid, &workspace.workspace_id)
-        .await?
-    );
-
-    // wait for update message
-    let mut workspace_listener = listeners.subscribe_workspace_member_change();
-
-    biz::workspace::ops::update_workspace_member(
-      &pool,
-      &workspace.workspace_id,
-      &WorkspaceMemberChangeset {
-        email: member.email.clone(),
-        role: Some(AFRole::Guest),
-        name: None,
-      },
-    )
-    .await
-    .context("update user workspace role")?;
-
-    let _ = workspace_listener.recv().await;
-
-    assert_eq!(
-      AFRole::Guest,
-      access_control
-        .get_role_from_uid(&member.uid, &workspace.workspace_id)
-        .await?
-    );
-
-    // wait for delete message
-    let mut workspace_listener = listeners.subscribe_workspace_member_change();
-
-    biz::workspace::ops::remove_workspace_members(
-      &user.uuid,
-      &pool,
-      &workspace.workspace_id,
-      &[member.email.clone()],
-    )
-    .await
-    .context("removing users from workspace")?;
-
-    let _ = workspace_listener.recv().await;
-
-    assert!(access_control
-      .get_role_from_uid(&member.uid, &workspace.workspace_id)
-      .await
-      .expect_err("user should not be part of workspace")
-      .is_not_enough_permissions());
-
-    Ok(())
-  }
-
-  #[sqlx::test(migrations = false)]
-  async fn test_collab_access_control_get_access_level(pool: PgPool) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-    let access_control = access_control.new_collab_access_control();
-
-    let user = tests::create_user(&pool).await?;
-    let owner = tests::create_user(&pool).await?;
-    let member = tests::create_user(&pool).await?;
-    let guest = tests::create_user(&pool).await?;
-
-    // Get workspace details
-    let workspace = database::workspace::select_user_workspace(&pool, &user.uuid)
-      .await?
-      .into_iter()
-      .next()
-      .ok_or(anyhow!("workspace should be created"))?;
-
-    let members = vec![
-      CreateWorkspaceMember {
-        email: owner.email.clone(),
-        role: AFRole::Owner,
-      },
-      CreateWorkspaceMember {
-        email: member.email.clone(),
-        role: AFRole::Member,
-      },
-      CreateWorkspaceMember {
-        email: guest.email.clone(),
-        role: AFRole::Guest,
-      },
-    ];
-    let _ = biz::workspace::ops::add_workspace_members(
-      &pool,
-      &user.uuid,
-      &workspace.workspace_id,
-      members,
-    )
-    .await
-    .context("adding users to workspace")?;
-
-    assert_eq!(
-      AFAccessLevel::FullAccess,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserUuid(&user.uuid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-    assert_eq!(
-      AFAccessLevel::FullAccess,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserId(&user.uid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-
-    assert_eq!(
-      AFAccessLevel::FullAccess,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserId(&owner.uid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-
-    assert_eq!(
-      AFAccessLevel::ReadAndWrite,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserId(&member.uid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-
-    assert_eq!(
-      AFAccessLevel::ReadOnly,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserId(&guest.uid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-
-    // wait for update message
-    let mut collab_listener = listeners.subscribe_collab_member_change();
-
-    let mut txn = pool
-      .begin()
-      .await
-      .context("acquire transaction to update collab member")?;
-
-    database::collab::upsert_collab_member_with_txn(
-      guest.uid,
-      &workspace.workspace_id.to_string(),
-      &AFAccessLevel::ReadAndComment,
-      &mut txn,
-    )
-    .await?;
-
-    txn
-      .commit()
-      .await
-      .expect("commit transaction to update collab member");
-
-    let _ = collab_listener.recv().await;
-
-    assert_eq!(
-      AFAccessLevel::ReadAndComment,
-      access_control
-        .get_collab_access_level(
-          CollabUserId::UserId(&guest.uid),
-          &workspace.workspace_id.to_string(),
-        )
-        .await?
-    );
-
-    // wait for delete message
-    let mut collab_listener = listeners.subscribe_collab_member_change();
-
-    database::collab::delete_collab_member(guest.uid, &workspace.workspace_id.to_string(), &pool)
-      .await
-      .context("delete collab member")?;
-
-    let _ = collab_listener.recv().await;
-
-    assert!(access_control
-      .get_collab_access_level(
-        CollabUserId::UserId(&guest.uid),
-        &workspace.workspace_id.to_string()
-      )
-      .await
-      .expect_err("user should not be part of collab")
-      .is_record_not_found());
-
-    Ok(())
-  }
-
-  #[sqlx::test(migrations = false)]
-  async fn test_collab_access_control_access_http_method(pool: PgPool) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-    let access_control = access_control.new_collab_access_control();
-
-    let user = tests::create_user(&pool).await?;
-    let guest = tests::create_user(&pool).await?;
-    let stranger = tests::create_user(&pool).await?;
-
-    // Get workspace details
-    let workspace = database::workspace::select_user_workspace(&pool, &user.uuid)
-      .await?
-      .into_iter()
-      .next()
-      .ok_or(anyhow!("workspace should be created"))?;
-
-    let _ = biz::workspace::ops::add_workspace_members(
-      &pool,
-      &guest.uuid,
-      &workspace.workspace_id,
-      vec![CreateWorkspaceMember {
-        email: guest.email,
-        role: AFRole::Guest,
-      }],
-    )
-    .await
-    .context("adding users to workspace")?;
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&user.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::GET
-        )
-        .await?
-    );
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&user.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::POST
-        )
-        .await?
-    );
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&user.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::PUT
-        )
-        .await?
-    );
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&user.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::DELETE
-        )
-        .await?
-    );
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&user.uid),
-          "new collab oid",
-          &Method::POST
-        )
-        .await?,
-      "should have access to non-existent collab oid"
-    );
-
-    assert!(
-      access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&guest.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::GET
-        )
-        .await?,
-      "guest should have read access"
-    );
-
-    assert!(
-      !access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&guest.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::POST
-        )
-        .await?,
-      "guest should not have write access"
-    );
-
-    assert!(
-      !access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&stranger.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::GET
-        )
-        .await?,
-      "stranger should not have read access"
-    );
-
-    assert!(
-      !access_control
-        .can_access_http_method(
-          CollabUserId::UserId(&stranger.uid),
-          &workspace.workspace_id.to_string(),
-          &Method::POST
-        )
-        .await?,
-      "stranger should not have write access"
-    );
-
-    Ok(())
-  }
-
-  #[sqlx::test(migrations = false)]
-  async fn test_collab_access_control_send_receive_collab_update(
-    pool: PgPool,
-  ) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-    let access_control = access_control.new_collab_access_control();
-
-    let user = tests::create_user(&pool).await?;
-    let guest = tests::create_user(&pool).await?;
-    let stranger = tests::create_user(&pool).await?;
-
-    // Get workspace details
-    let workspace = database::workspace::select_user_workspace(&pool, &user.uuid)
-      .await?
-      .into_iter()
-      .next()
-      .ok_or(anyhow!("workspace should be created"))?;
-
-    let _ = biz::workspace::ops::add_workspace_members(
-      &pool,
-      &guest.uuid,
-      &workspace.workspace_id,
-      vec![CreateWorkspaceMember {
-        email: guest.email,
-        role: AFRole::Guest,
-      }],
-    )
-    .await
-    .context("adding users to workspace")?;
-
-    assert!(
-      access_control
-        .can_send_collab_update(&user.uid, &workspace.workspace_id.to_string())
-        .await?
-    );
-
-    assert!(
-      access_control
-        .can_receive_collab_update(&user.uid, &workspace.workspace_id.to_string())
-        .await?
-    );
-
-    assert!(
-      !access_control
-        .can_send_collab_update(&guest.uid, &workspace.workspace_id.to_string())
-        .await?,
-      "guest cannot send collab update"
-    );
-
-    assert!(
-      access_control
-        .can_receive_collab_update(&guest.uid, &workspace.workspace_id.to_string())
-        .await?,
-      "guest can receive collab update"
-    );
-
-    assert!(
-      !access_control
-        .can_send_collab_update(&stranger.uid, &workspace.workspace_id.to_string())
-        .await?,
-      "stranger cannot send collab update"
-    );
-
-    assert!(
-      !access_control
-        .can_receive_collab_update(&stranger.uid, &workspace.workspace_id.to_string())
-        .await?,
-      "stranger cannot receive collab update"
-    );
-
-    Ok(())
-  }
-
-  #[sqlx::test(migrations = false)]
-  async fn test_collab_access_control_cache_collab_access_level(
-    pool: PgPool,
-  ) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-    let access_control = access_control.new_collab_access_control();
-
-    let uid = 123;
-    let oid = "collab::oid".to_owned();
-    access_control
-      .cache_collab_access_level(CollabUserId::UserId(&uid), &oid, AFAccessLevel::FullAccess)
-      .await?;
-
-    assert_eq!(
-      AFAccessLevel::FullAccess,
-      access_control
-        .get_collab_access_level(CollabUserId::UserId(&uid), &oid)
-        .await?
-    );
-
-    access_control
-      .cache_collab_access_level(CollabUserId::UserId(&uid), &oid, AFAccessLevel::ReadOnly)
-      .await?;
-
-    assert_eq!(
-      AFAccessLevel::ReadOnly,
-      access_control
-        .get_collab_access_level(CollabUserId::UserId(&uid), &oid)
-        .await?
-    );
-
-    Ok(())
-  }
-  #[sqlx::test(migrations = false)]
-  async fn test_casbin_access_control_update_remove(pool: PgPool) -> anyhow::Result<()> {
-    tests::setup_db(&pool).await?;
-
-    let model = DefaultModel::from_str(tests::MODEL_CONF).await?;
-    let enforcer = Enforcer::new(model, PgAdapter::new(pool.clone())).await?;
-    let listeners = PgListeners::new(&pool).await?;
-    let access_control = CasbinAccessControl::new(
-      pool.clone(),
-      listeners.subscribe_collab_member_change(),
-      listeners.subscribe_workspace_member_change(),
-      enforcer,
-    );
-
-    let uid = 123;
-    assert!(
-      access_control
-        .update(
-          &uid,
-          &ObjectType::Workspace("123"),
-          &ActionType::Role(AFRole::Owner)
-        )
-        .await?
-    );
-
-    assert!(access_control.enforcer.read().await.enforce((
-      uid.to_string(),
-      ObjectType::Workspace("123").to_string(),
-      i32::from(AFRole::Owner).to_string(),
-    ))?);
-
-    assert!(
-      access_control
-        .remove(&uid, &ObjectType::Workspace("123"))
-        .await?
-    );
-
-    assert!(!access_control.enforcer.read().await.enforce((
-      uid.to_string(),
-      ObjectType::Workspace("123").to_string(),
-      i32::from(AFRole::Owner).to_string(),
-    ))?);
-
     Ok(())
   }
 }
