+++ src/biz/mod.rs
@@ -3,4 +3,4 @@ pub mod collab;
 pub mod pg_listener;
 pub mod user;
 pub mod utils;
-pub(crate) mod workspace;
+pub mod workspace;
